# IoTApi/config/swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "v1",
		"title": "IoT API",
		"description": "APIs for Azure IoT POC"
	},
	"securityDefinitions": {
		"Authorization": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header",
			"description": "SAS Token"
		}
	},
	"paths": {
		"/generateSasToken": {
			"post": {
				"description": "Generate SAS Token as Authorization for other APIs",
				"tags": [
					"Token"
				],
				"operationId": "generateSasToken",
				"consumes": [],
				"produces": [
					"application/json",
					"text/json"
				],
				"parameters": [
					{
						"name": "SasRequest",
						"in": "body",
						"schema": {
							"$ref": "#/definitions/SasRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/SasToken"
						}
					}
				},
				"deprecated": false
			}
		},
		"/IoTEvents/{deviceId}": {
			"get": {
				"description": "Retrieve device's events with the deviceId",
				"tags": [
					"Events"
				],
				"operationId": "getIoTEvents",
				"consumes": [],
				"produces": [
					"application/json",
					"text/json"
				],
				"parameters": [
					{
						"name": "deviceId",
						"in": "path",
						"description": "Id of the Device for retrieving the Events",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/GPSEvent"
							}
						}
					}
				},
				"deprecated": false
			}
		},
		"/IoTEvents/search": {
			"post": {
				"description": "Search IoT Events",
				"tags": [
					"Events"
				],
				"operationId": "searchIoTEvent",
				"consumes": [
					"application/json",
					"text/json"
				],
				"produces": [
					"application/json",
					"text/json"
				],
				"parameters": [
					{
						"name": "searchCondition",
						"in": "body",
						"schema": {
							"$ref": "#/definitions/IoTEventSearchCondition"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/GPSEvent"
							}
						}
					}
				},
				"deprecated": false
			}
		},
		"/IoTDevices": {
			"post": {
				"description": "Register the device to IoTHub and record custom meta data in DocumentDB",
				"tags": [
					"Devices"
				],
				"operationId": "registerIoTDevice",
				"consumes": [
					"application/json",
					"text/json"
				],
				"produces": [
					"application/json",
					"text/json"
				],
				"parameters": [
					{
						"name": "IoTDevice",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IoTDeviceCreation"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/IoTDevice"
						}
					}
				},
				"deprecated": false
			},
			"put": {
				"description": "Update the device custom meta data",
				"tags": [
					"Devices"
				],
				"operationId": "registerIoTDevice",
				"consumes": [
					"application/json",
					"text/json"
				],
				"produces": [
					"application/json",
					"text/json"
				],
				"parameters": [
					{
						"name": "IoTDevice",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IoTDeviceUpdate"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/IoTDevice"
						}
					}
				},
				"deprecated": false
			}
		},
		"/IoTDevices/search": {
			"post": {
				"description": "Search IoT Devices with meta data key/values",
				"tags": [
					"Devices"
				],
				"operationId": "searchIoTDevice",
				"consumes": [
					"application/json",
					"text/json"
				],
				"produces": [
					"application/json",
					"text/json"
				],
				"parameters": [
					{
						"name": "searchCondition",
						"in": "body",
						"schema": {
							"$ref": "#/definitions/IoTDeviceSearchCondition"
						}
					},
					{
						"name": "showKeys",
						"in": "query",
						"type": "boolean",
						"default": false
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/IoTDevice"
							}
						}
					}
				},
				"deprecated": false
			}
		},
		"/IoTDevices/search/activation": {
			"post": {
				"description": "Search IoT Devices with activation code",
				"tags": [
					"Devices"
				],
				"operationId": "searchIoTDeviceByActivationCode",
				"consumes": [
					"application/json",
					"text/json"
				],
				"produces": [
					"application/json",
					"text/json"
				],
				"parameters": [
					{
						"name": "activationCode",
						"in": "body",
						"schema": {
							"type": "object",
							"properties": {
								"activationCode": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/IoTDevice"
						}
					}
				},
				"deprecated": false
			}
		},
		"/IoTDevices/{deviceId}": {
			"get": {
				"description": "Retrieve device's information with the deviceId",
				"tags": [
					"Devices"
				],
				"operationId": "getIoTDevice",
				"consumes": [
					"application/json",
					"text/json"
				],
				"produces": [
					"application/json",
					"text/json"
				],
				"parameters": [
					{
						"name": "deviceId",
						"in": "path",
						"description": "Id of the Device",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/IoTDevice"
						}
					}
				},
				"deprecated": false
			}
		}
	},
	"definitions": {
		"SasRequest": {
			"type": "object",
			"properties": {
				"resourceUri": {
					"type": "string"
				},
				"signingKey": {
					"type": "string"
				},
				"policyName": {
					"type": "string"
				},
				"expiresInMins": {
					"format": "int32",
					"type": "integer",
					"default": 15
				}
			}
		},
		"SasToken": {
			"type": "object",
			"properties": {
				"token": {
					"type": "string"
				}
			}
		},
		"IoTDevice": {
			"type": "object",
			"properties": {
				"deviceId": {
					"type": "string"
				},
				"generationId": {
					"type": "string"
				},
				"activationCode": {
					"type": "string",
					"description": "Unique activation code to activate the app"
				},
				"etag": {
					"type": "string"
				},
				"connectionState": {
					"type": "string"
				},
				"status": {
					"type": "string"
				},
				"statusReason": {
					"type": "string"
				},
				"statusUpdateTime": {
					"type": "string",
					"format": "date-time"
				},
				"lastActivityTime": {
					"type": "string",
					"format": "date-time"
				},
				"authentication": {
					"type": "object",
					"properties": {
						"symmetricKey": {
							"type": "object",
							"properties": {
								"primaryKey": {
									"type": "string"
								},
								"secondaryKey": {
									"type": "string"
								}
							}
						},
						"x509Thumbprint": {
							"type": "object",
							"properties": {
								"primaryThumbprint": {
									"type": "string"
								},
								"secondaryThumbprint": {
									"type": "string"
								}
							}
						}
					}
				},
				"meta": {
					"type": "array",
					"items": {
						"type": "object",
						"properties": {
							"key": {
								"type": "string",
								"description": "Name of the Meta Field"
							},
							"value": {
								"type": "string",
								"description": "Value of the Meta Field"
							}
						}
					}
				}
			}
		},
		"IoTDeviceSearchCondition": {
			"type": "array",
			"items": {
				"type": "object",
				"required": ["key", "value"],
				"properties": {
					"key": {
						"type": "string",
						"description": "Name of the Meta Field"
					},
					"value": {
						"type": "string",
						"description": "Value of the Meta Field"
					}
				}
			}
		},
		"IoTEventSearchCondition": {
			"type": "array",
			"items": {
				"type": "object",
				"required": ["key", "value"],
				"properties": {
					"key": {
						"type": "string",
						"description": "Name of the Event Field"
					},
					"operator": {
						"type": "string",
						"description": "Operator to comparing the field value, currently equal only",
						"enum": ["="],
						"default": "="
					},
					"value": {
						"type": "string",
						"description": "Value of the Event Field"
					}
				}
			}
		},
		"IoTDeviceCreation": {
			"type": "object",
			"required": ["activationCode"],
			"properties": {
				"deviceId": {
					"type": "string",
					"description": "Device Id will be auto-generated if this field is empty"
				},
				"activationCode": {
					"type": "string",
					"description": "Unique activation code to activate the app"
				},
				"connectionState": {
					"type": "string",
					"enum": ["connected", "disconnected"],
					"default": "connected"
				},
				"status": {
					"type": "string",
					"enum": ["enabled", "disabled"],
					"default": "enabled"
				},
				"statusReason": {
					"type": "string"
				},
				"connectionStateUpdatedTime": {
					"type": "string",
					"format": "date-time"
				},
				"statusUpdatedTime": {
					"type": "string",
					"format": "date-time"
				},
				"lastActivityTime": {
					"type": "string",
					"format": "date-time"
				},
				"meta": {
					"type": "array",
					"items": {
						"type": "object",
						"properties": {
							"key": {
								"type": "string",
								"description": "Name of the Meta Field"
							},
							"value": {
								"type": "string",
								"description": "Value of the Meta Field"
							}
						}
					}
				}
			}
		},
		"IoTDeviceUpdate": {
			"type": "object",
			"properties": {
				"deviceId": {
					"type": "string",
					"description": "Device Id will be auto-generated if this field is empty"
				},
				"meta": {
					"type": "array",
					"items": {
						"type": "object",
						"properties": {
							"key": {
								"type": "string",
								"description": "Name of the Meta Field"
							},
							"value": {
								"type": "string",
								"description": "Value of the Meta Field"
							}
						}
					}
				}
			}
		},
		"VendorIoTDevice": {
			"type": "object",
			"properties": {
				"deviceId": {
					"type": "string"
				},
				"etag": {
					"type": "string"
				},
				"lastUpdated": {
					"format": "date-time",
					"type": "string"
				},
				"meta": {
					"type": "array",
					"items": {
						"type": "object",
						"properties": {
							"key": {
								"type": "string",
								"description": "Name of the Meta Field"
							},
							"value": {
								"type": "string",
								"description": "Value of the Meta Field"
							}
						}
					}
				}
			}
		},
		"GPSEvent": {
			"type": "object",
			"properties": {
				"routeID": {
					"type": "string"
				},
				"routeFlag": {
					"type": "string"
				},
				"longitude": {
					"format": "double",
					"type": "number"
				},
				"latitude": {
					"format": "double",
					"type": "number"
				},
				"altitude": {
					"format": "double",
					"type": "number"
				},
				"hAccuracy": {
					"format": "double",
					"type": "number"
				},
				"vAccuracy": {
					"format": "double",
					"type": "number"
				},
				"speed": {
					"format": "double",
					"type": "number"
				},
				"angle": {
					"format": "double",
					"type": "number"
				},
				"distance": {
					"format": "double",
					"type": "number"
				},
				"cordinateSystem": {
					"type": "string"
				},
				"timeStamp": {
					"format": "date-time",
					"type": "string"
				},
				"timezone": {
					"type": "string"
				},
				"devicePower": {
					"format": "double",
					"type": "number"
				},
				"intervalTime": {
					"format": "int32",
					"type": "number"
				},
				"intervalDistance": {
					"format": "double",
					"type": "number"
				}
			}
		},
		"IoTEvent": {
			"type": "object",
			"properties": {
				"deviceId": {
					"type": "string"
				},
				"windSpeed": {
					"format": "double",
					"type": "number"
				},
				"EventProcessedUtcTime": {
					"format": "date-time",
					"type": "string"
				},
				"IoTHub": {
					"$ref": "#/definitions/IoTHub"
				}
			}
		},
		"IoTHub": {
			"type": "object",
			"properties": {
				"MessageId": {
					"type": "string"
				},
				"CorrelationId": {
					"type": "string"
				}
			}
		}
	}
}